Extract hook useElementWidth
